fixed makesym syntax error

+++ .makesym.sh
@@ -7,7 +7,6 @@ olddir=~/.dotfiles_old
 files="bashrc bash_profile vim dircolors gitconfig Xresources config xinitrc tmux.conf dircolors bw-scripts do-scripts chunkwm_plugins chunkwmrc"
 
 # Create .dotfiles_old in homedir
-echo -n "Creating $olddir for backup of any existing dotfiles..."
 mkdir -p $olddir
 echo "done"
 
@@ -17,6 +16,9 @@ cd $dir
 # Move any existing dotfiles to dotfiles_old directory, and create symlinks
 # from the homedir to any files in the ~/dotfiles
 for file in $files; do
-    mv ~/.$file ~/.dotfiles_old/
+    if [[ -f ~/.file ]]
+    then
+        mv ~/.$file ~/.dotfiles_old/
+    fi
     ln -s $dir/$file ~/.$file
 done
